requirements.txt
Flask==2.0.3
numpy==1.23.2
Pillow==9.0.0
torch==1.12.1+cu116
torchaudio==0.12.1+cu116
torchvision==0.13.1+cu116
opencv-contrib-python==4.5.5.64
Flask-Cors==3.0.10